+++ core-plugins/src/test/java/co/cask/hydrator/plugin/transform/PythonEvaluatorTest.java
@@ -96,8 +96,8 @@ public class PythonEvaluatorTest {
     // check record1
     Assert.assertEquals(SCHEMA, output.getSchema());
     Assert.assertTrue((Boolean) output.get("booleanField"));
-    Assert.assertEquals(28 * 1024, output.get("intField"));
-    Assert.assertEquals(99L, output.get("longField"));
+    Assert.assertEquals(28 * 1024, (int) output.get("intField"));
+    Assert.assertEquals(99L, (long) output.get("longField"));
     Assert.assertTrue(Math.abs(2.71f - (Float) output.get("floatField")) < 0.000001);
     Assert.assertTrue(Math.abs(3.14 - (Double) output.get("doubleField")) < 0.000001);
     Assert.assertArrayEquals(Bytes.toBytes("foo"), (byte[]) output.get("bytesField"));
@@ -119,14 +119,14 @@ public class PythonEvaluatorTest {
 
     Assert.assertEquals(SCHEMA, output.getSchema());
     Assert.assertFalse((Boolean) output.get("booleanField"));
-    Assert.assertEquals(-28 * 1024, output.get("intField"));
-    Assert.assertEquals(-99L, output.get("longField"));
+    Assert.assertEquals(-28 * 1024, (int) output.get("intField"));
+    Assert.assertEquals(-99L, (long) output.get("longField"));
     Assert.assertTrue(Math.abs(-2.71f - (Float) output.get("floatField")) < 0.000001);
     Assert.assertTrue(Math.abs(-3.14 - (Double) output.get("doubleField")) < 0.000001);
     Assert.assertArrayEquals(Bytes.toBytes("hello"), (byte[]) output.get("bytesField"));
     Assert.assertEquals("world", output.get("stringField"));
     Assert.assertNull(output.get("nullableField"));
-    Assert.assertEquals(3, output.get("unionField"));
+    Assert.assertEquals(3, (int) output.get("unionField"));
     expectedMapField = ImmutableMap.of();
     expectedListField = ImmutableList.of();
     Assert.assertEquals(expectedMapField, output.get("mapField"));
@@ -151,7 +151,7 @@ public class PythonEvaluatorTest {
 
     // check simple types are decoded properly
     Assert.assertEquals(SCHEMA, output.getSchema());
-    Assert.assertEquals(198L, output.get("longField"));
+    Assert.assertEquals(198L, (long) output.get("longField"));
     Assert.assertTrue(Math.abs(2.71f - (Float) output.get("floatField")) < 0.03);
     Assert.assertTrue(Math.abs(3.14 - (Double) output.get("doubleField")) < 0.032);
     emitter.clear();
@@ -251,8 +251,8 @@ public class PythonEvaluatorTest {
     transform.transform(RECORD1, emitter);
     StructuredRecord output = emitter.getEmitted().get(0);
     Assert.assertEquals(outputSchema, output.getSchema());
-    Assert.assertEquals(28, output.get("x"));
-    Assert.assertEquals(99L, output.get("y"));
+    Assert.assertEquals(28, (int) output.get("x"));
+    Assert.assertEquals(99L, (long) output.get("y"));
   }
 
   @Test
